% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{combine}
\alias{combine}
\title{functions to combine raster results
into new raster}
\usage{
combine(name, raster, wsf)
}
\arguments{
\item{wsf_raster}{raster with world settlement footprint data}

\item{boundary_polygon}{vector file with optional boundary units}
}
\value{
total built-up area as sf-object
}
\description{
functions to combine raster results
into new raster
}
