% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_results.R
\name{combine_results}
\alias{combine_results}
\title{functions to combine results extracted from the rasters
into data frame and and export as csv if needed}
\usage{
combine_results(name, raster, wsf, export = FALSE)
}
\arguments{
\item{name}{vector with objective names}

\item{raster}{vector extracted mean raster values}

\item{wsf}{vector with extracted wsf values}

\item{export}{boolean if an export of the matrix in csv-format is needed}
}
\value{
data frame with name columnn and extracted values of examined raster
and wsf
}
\description{
functions to combine results extracted from the rasters
into data frame and and export as csv if needed
}
