% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_example_data.R
\name{load_example_data}
\alias{load_example_data}
\title{function to load examplary land surface temperature data for Berlin}
\usage{
load_example_data()
}
\value{
three variables with district polygons, wsf raster
and summer lst mean raster
}
\description{
function to load examplary land surface temperature data for Berlin
}
