% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rastmeancalc.R
\name{rastmeancalc}
\alias{rastmeancalc}
\title{function to calculate mean value for a raster
based on boundary polygon}
\usage{
rastmeancalc(raster, boundary_polygon)
}
\arguments{
\item{raster}{raster with random data}

\item{boundary_polygon}{vector file with boundary units}
}
\value{
mean value for raster values
}
\description{
function to calculate mean value for a raster
based on boundary polygon
}
