% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_correlation.R
\name{visual_correlation}
\alias{visual_correlation}
\title{function to calculate the correlation between world settlement footprint and
the examined variable and visualisation in a graph}
\usage{
visual_correlation(results, y_axis = NULL)
}
\arguments{
\item{results}{dataframe with three columns, return result of combine-function}

\item{y_axis}{name of the y-axis that should be displayed}
}
\value{
a gg2-plot graph with the results
}
\description{
function to calculate the correlation between world settlement footprint and
the examined variable and visualisation in a graph
}
