% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsfcalc.R
\name{wsfcalc}
\alias{wsfcalc}
\title{wsfcalc()}
\usage{
wsfcalc(wsf_raster, boundary_polygon)
}
\arguments{
\item{wsf_raster}{raster with world settlement footprint}

\item{boundary_polygon}{vector file with boundary units}
}
\value{
mean value for world settlement footprint in \%
}
\description{
function to calculate mean wsf value for a raster
based on boundary polygon
}
